moved LandingPage to components directory
